Add MongoDB and add sample code to how it works
I added the functionality to how it works in the mongoTester class.
Check the main function, thats where it gets called. I commented out
the run function so the server won’t actually do anything but test the
mongo db. My example should cover most the stuff we need for the DB